refactor: define interface in consumer
